API_dev/Tutorial_0/source/api.backend/requirements.txt
requests
fastapi
ujson
uvicorn
mysql-connector-python
